+++ premium/backend/src/baserow_premium/views/decorator_value_provider_types.py
@@ -1,22 +1,53 @@
+from typing import Any, Dict
+
 from baserow.contrib.database.fields.field_types import (
     SingleSelectFieldType,
 )
-
-from baserow.contrib.database.views.models import ViewDecoration
+from baserow.contrib.database.views.models import (
+    ViewDecoration,
+)
 from baserow.contrib.database.views.handler import ViewHandler
 from baserow.contrib.database.views.registries import (
     DecoratorValueProviderType,
     view_filter_type_registry,
 )
 
+from baserow_premium.license.handler import check_active_premium_license
 
-class SelectColorValueProviderType(DecoratorValueProviderType):
+from .decorator_types import BackgroundColorDecoratorType, LeftBorderColorDecoratorType
+from .serializers import (
+    SelectColorValueProviderConfSerializer,
+    ConditionalColorValueProviderConfColorsSerializer,
+)
+
+
+class PremiumDecoratorValueProviderType(DecoratorValueProviderType):
+    def before_create_decoration(self, view, user):
+        if user:
+            check_active_premium_license(user)
+
+    def before_update_decoration(self, view, user):
+        if user:
+            check_active_premium_license(user)
+
+
+class SelectColorValueProviderType(PremiumDecoratorValueProviderType):
     type = "single_select_color"
 
-    def set_import_serialized_value(self, value, id_mapping) -> str:
+    compatible_decorator_types = [
+        LeftBorderColorDecoratorType.type,
+        BackgroundColorDecoratorType.type,
+    ]
+
+    value_provider_conf_serializer_class = SelectColorValueProviderConfSerializer
+
+    def set_import_serialized_value(
+        self, value: Dict[str, Any], id_mapping: Dict[str, Any]
+    ) -> Dict[str, Any]:
         """
         Update the field id with the newly created one.
         """
+
         old_field_id = value["value_provider_conf"].get("field_id", None)
 
         if old_field_id:
@@ -69,10 +100,21 @@ class SelectColorValueProviderType(DecoratorValueProviderType):
                 view_handler.update_decoration(decoration, value_provider_conf=new_conf)
 
 
-class ConditionalColorValueProviderType(DecoratorValueProviderType):
+class ConditionalColorValueProviderType(PremiumDecoratorValueProviderType):
     type = "conditional_color"
 
-    def set_import_serialized_value(self, value, id_mapping) -> str:
+    compatible_decorator_types = [
+        LeftBorderColorDecoratorType.type,
+        BackgroundColorDecoratorType.type,
+    ]
+
+    value_provider_conf_serializer_class = (
+        ConditionalColorValueProviderConfColorsSerializer
+    )
+
+    def set_import_serialized_value(
+        self, value: Dict[str, Any], id_mapping: Dict[str, Any]
+    ) -> Dict[str, Any]:
         """
         Update the field ids of each filter with the newly created one.
         """