+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Fixed translations in emails sent by Baserow.
 * Fixed invalid `first_name` validation in the account form modal.
 * Shared public forms now don't allow creating new options
   for single and multiple select fields.
@@ -18,6 +19,7 @@
 * Added `is days ago` filter to date field.
 * Fixed a bug that made it possible to delete created on/modified by fields on the web frontend.
 * Allow the setting of max request page size via environment variable.
+* Added select option suggestions when converting to a select field.
 * Introduced read only lookup of foreign row by clicking on a link row relationship in 
   the grid view row modal.
 * Boolean field converts the word `checked` to `True` value.
@@ -44,6 +46,8 @@
   requests are made.
 * Added password protection for publicly shared grids and forms.
 * Made views trashable.
+* Fixed bug where a cell value was not reverted when the request to the backend fails.
+* **Premium** Added row coloring.
 
 ## Released (2022-03-03 1.9.1)